+++ map_iter.go
@@ -1,0 +1,202 @@
+package btree
+
+// MapIter represents an iterator for btree.Map
+type MapIter[K ordered, V any] struct {
+	tr      *Map[K, V]
+	mut     bool
+	seeked  bool
+	atstart bool
+	atend   bool
+	stack   []mapIterStackItem[K, V]
+	item    mapPair[K, V]
+}
+
+type mapIterStackItem[K ordered, V any] struct {
+	n *mapNode[K, V]
+	i int
+}
+
+// Seek to item greater-or-equal-to key.
+// Returns false if there was no item found.
+func (iter *MapIter[K, V]) Seek(key K) bool {
+	if iter.tr == nil {
+		return false
+	}
+	iter.seeked = true
+	iter.stack = iter.stack[:0]
+	if iter.tr.root == nil {
+		return false
+	}
+	n := iter.tr.isoLoad(&iter.tr.root, iter.mut)
+	for {
+		i, found := iter.tr.search(n, key)
+		iter.stack = append(iter.stack, mapIterStackItem[K, V]{n, i})
+		if found {
+			iter.item = n.items[i]
+			return true
+		}
+		if n.leaf() {
+			iter.stack[len(iter.stack)-1].i--
+			return iter.Next()
+		}
+		n = iter.tr.isoLoad(&(*n.children)[i], iter.mut)
+	}
+}
+
+// First moves iterator to first item in tree.
+// Returns false if the tree is empty.
+func (iter *MapIter[K, V]) First() bool {
+	if iter.tr == nil {
+		return false
+	}
+	iter.atend = false
+	iter.atstart = false
+	iter.seeked = true
+	iter.stack = iter.stack[:0]
+	if iter.tr.root == nil {
+		return false
+	}
+	n := iter.tr.isoLoad(&iter.tr.root, iter.mut)
+	for {
+		iter.stack = append(iter.stack, mapIterStackItem[K, V]{n, 0})
+		if n.leaf() {
+			break
+		}
+		n = iter.tr.isoLoad(&(*n.children)[0], iter.mut)
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Last moves iterator to last item in tree.
+// Returns false if the tree is empty.
+func (iter *MapIter[K, V]) Last() bool {
+	if iter.tr == nil {
+		return false
+	}
+	iter.seeked = true
+	iter.stack = iter.stack[:0]
+	if iter.tr.root == nil {
+		return false
+	}
+	n := iter.tr.isoLoad(&iter.tr.root, iter.mut)
+	for {
+		iter.stack = append(iter.stack, mapIterStackItem[K, V]{n, len(n.items)})
+		if n.leaf() {
+			iter.stack[len(iter.stack)-1].i--
+			break
+		}
+		n = iter.tr.isoLoad(&(*n.children)[len(n.items)], iter.mut)
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Next moves iterator to the next item in iterator.
+// Returns false if the tree is empty or the iterator is at the end of
+// the tree.
+func (iter *MapIter[K, V]) Next() bool {
+	if iter.tr == nil {
+		return false
+	}
+	if !iter.seeked {
+		return iter.First()
+	}
+	if len(iter.stack) == 0 {
+		if iter.atstart {
+			return iter.First() && iter.Next()
+		}
+		return false
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	s.i++
+	if s.n.leaf() {
+		if s.i == len(s.n.items) {
+			for {
+				iter.stack = iter.stack[:len(iter.stack)-1]
+				if len(iter.stack) == 0 {
+					iter.atend = true
+					return false
+				}
+				s = &iter.stack[len(iter.stack)-1]
+				if s.i < len(s.n.items) {
+					break
+				}
+			}
+		}
+	} else {
+		n := iter.tr.isoLoad(&(*s.n.children)[s.i], iter.mut)
+		for {
+			iter.stack = append(iter.stack, mapIterStackItem[K, V]{n, 0})
+			if n.leaf() {
+				break
+			}
+			n = iter.tr.isoLoad(&(*n.children)[0], iter.mut)
+		}
+	}
+	s = &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Prev moves iterator to the previous item in iterator.
+// Returns false if the tree is empty or the iterator is at the beginning of
+// the tree.
+func (iter *MapIter[K, V]) Prev() bool {
+	if iter.tr == nil {
+		return false
+	}
+	if !iter.seeked {
+		return false
+	}
+	if len(iter.stack) == 0 {
+		if iter.atend {
+			return iter.Last() && iter.Prev()
+		}
+		return false
+	}
+	s := &iter.stack[len(iter.stack)-1]
+	if s.n.leaf() {
+		s.i--
+		if s.i == -1 {
+			for {
+				iter.stack = iter.stack[:len(iter.stack)-1]
+				if len(iter.stack) == 0 {
+					iter.atstart = true
+					return false
+				}
+				s = &iter.stack[len(iter.stack)-1]
+				s.i--
+				if s.i > -1 {
+					break
+				}
+			}
+		}
+	} else {
+		n := iter.tr.isoLoad(&(*s.n.children)[s.i], iter.mut)
+		for {
+			iter.stack = append(iter.stack,
+				mapIterStackItem[K, V]{n, len(n.items)})
+			if n.leaf() {
+				iter.stack[len(iter.stack)-1].i--
+				break
+			}
+			n = iter.tr.isoLoad(&(*n.children)[len(n.items)], iter.mut)
+		}
+	}
+	s = &iter.stack[len(iter.stack)-1]
+	iter.item = s.n.items[s.i]
+	return true
+}
+
+// Key returns the current iterator item key.
+func (iter *MapIter[K, V]) Key() K {
+	return iter.item.key
+}
+
+// Value returns the current iterator item value.
+func (iter *MapIter[K, V]) Value() V {
+	return iter.item.value
+}
